backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
bleach==6.1.0
python-decouple==3.8
Pillow==10.4.0
setuptools>=65.0.0

# PostgreSQL
psycopg2-binary==2.9.9

# Redis (optional - for future)
redis==5.0.1
django-redis==5.4.0

# Database URL helper
dj-database-url==2.1.0

# Production server
gunicorn==21.2.0

# Static files management
whitenoise==6.6.0

# AWS SDK (for S3 in future)
boto3==1.34.34

# Health checks
django-health-check==3.18.1
